- fix bug 870794. Direct messages between 64/32-bit machines did not work properly due to different length of "unsigned long" on 32/64-bit machines (used for message sequence numbers). The ICQ protocol does only use 2-byte sequence numbers, but licq did use unsigned long (4/8-byte) internally. One result of this bug was, that ACK packets could not be associated to the corresponding event. The fix is to use "unsigned short" for sequence numbers internally.

+++ licq/include/licq_chat.h
@@ -176,7 +176,7 @@ struct SChatReverseConnectInfo
 class CPacketChat : public CPacket
 {
 public:
-  virtual const unsigned long  Sequence()   { return 0; };
+  virtual const unsigned short Sequence()   { return 0; };
   virtual const unsigned short SubSequence()   { return 0; };
   virtual const unsigned short Command()    { return 0; };
   virtual const unsigned short SubCommand() { return 0; };
