+++ licq/include/licq_filetransfer.h
@@ -87,10 +87,10 @@ struct SFileReverseConnectInfo
 class CPacketFile : public CPacket
 {
 public:
-  virtual const unsigned long  Sequence()   { return 0; };
-  virtual const unsigned short SubSequence()   { return 0; };
-  virtual const unsigned short Command()    { return 0; };
-  virtual const unsigned short SubCommand() { return 0; };
+  virtual const unsigned short Sequence()    { return 0; };
+  virtual const unsigned short SubSequence() { return 0; };
+  virtual const unsigned short Command()     { return 0; };
+  virtual const unsigned short SubCommand()  { return 0; };
 protected:
    void InitBuffer()   { buffer = new CBuffer(m_nSize); };
 };
